user_crud/pyproject.toml
[tool.poetry]
name = "user_crud"
version = "0.1.0"
description = ""
authors = ["Alan Bandeira <alan.p.bandeira@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
pydantic = {extras = ["email"], version = "^1.8.1"}
motor = "^2.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-dotenv = "^0.15.0"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
